server/requirements.txt
# Core
Flask==3.1.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
flask-sqlalchemy==3.1.1
alembic==1.13.1

# Search
elasticsearch==8.12.0
elasticsearch-dsl==8.12.0

# Caching
redis==5.0.1

# Text Processing
python-Levenshtein==0.23.0
unidecode==1.3.8

# API
flask-cors==4.0.0
flask-limiter==3.5.0

# Deployment
gunicorn==21.2.0
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3
